requirements.txt
annoy==1.17.3
google.generativeai==0.5.4
pandas==2.2.2
pyarrow==16.1.0
python-dotenv==1.0.1
sentence-transformers==3.0.0